# openapi.yaml
openapi: 3.1.0
info:
  title: Collider Custody - Transaction Security Layer
  description: |
    On-prem Transaction Security Layer + Wallet-as-a-Service (Ethereum)
    
    ## Features
    - Wallet Registry (WaaS) - Create and manage Ethereum wallets
    - Transaction Orchestrator - State machine for transaction lifecycle
    - KYT (Know Your Transaction) - Screen transactions
    - Policy Engine - Enforce limits and rules
    - Approvals with SoD - Multi-approval workflows
    - Signing Adapter - Dev mode signer
    - Chain Listener - Monitor confirmations
    - Audit Log - Tamper-evident hash-chain
  version: 1.0.0
  contact:
    name: Collider Team
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
      description: Request correlation ID for tracing
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Idempotency key for POST requests
  schemas:
...
    TxType:
      type: string
      enum: [WITHDRAW, TRANSFER, CONTRACT_CALL]
    TxStatus:
      type: string
      enum:
        - SUBMITTED
        - KYT_PENDING
        - KYT_BLOCKED
        - KYT_REVIEW
        - POLICY_EVAL_PENDING
        - POLICY_BLOCKED
        - APPROVAL_PENDING
        - REJECTED
        - SIGN_PENDING
        - SIGNED
        - FAILED_SIGN
        - BROADCAST_PENDING
        - BROADCASTED
        - FAILED_BROADCAST
        - CONFIRMING
        - CONFIRMED
        - FINALIZED
...
    UserRole:
      type: string
      enum: [ADMIN, OPERATOR, COMPLIANCE, VIEWER]
...
    TxRequestCreate:
      type: object
      required:
        - wallet_id
        - tx_type
        - to_address
        - amount
      properties:
        wallet_id:
          type: string
          format: uuid
        tx_type:
          $ref: '#/components/schemas/TxType'
        to_address:
          type: string
          pattern: ^0x[a-fA-F0-9]{40}$
        asset:
          type: string
          default: ETH
        amount:
          type: string
          format: decimal
        data:
          type: string
          description: Hex-encoded data for CONTRACT_CALL
...
    ApprovalCreate:
      type: object
      required:
        - decision
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED]
        comment:
          type: string
...
    KYTCaseResolve:
      type: object
      required:
        - decision
      properties:
        decision:
          type: string
          enum: [ALLOW, BLOCK]
        comment:
          type: string
...
    AuditPackageResponse:
      type: object
      properties:
        tx_request_id:
          type: string
        tx_request:
          type: object
        policy_evaluation:
          type: object
        kyt_evaluation:
          type: object
        approvals:
          type: array
          items:
            type: object
        signing:
          type: object
        broadcast:
          type: object
        confirmations:
          type: object
        audit_events:
          type: array
          items:
            type: object
        package_hash:
          type: string
        generated_at:
          type: string
          format: date-time
    AuditVerifyResponse:
      type: object
      properties:
        is_valid:
          type: boolean
        total_events:
          type: integer
        verified_events:
          type: integer
        first_event_id:
          type: string
        last_event_id:
          type: string
        chain_intact:
          type: boolean
        errors:
          type: array
          items:
            type: string
    CorrelatedResponse:
      type: object
      properties:
        correlation_id:
          type: string
        data:
          type: object
        timestamp:
          type: string
          format: date-time
...
paths:
  /v1/auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                role:
                  $ref: '#/components/schemas/UserRole'
      responses:
        '200':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrelatedResponse'
  /v1/auth/login:
    post:
      tags: [Authentication]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrelatedResponse'
...
  /v1/tx-requests:
    post:
      tags: [Transactions]
      summary: Create transaction request
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TxRequestCreate'
      responses:
        '200':
          description: Transaction request created
    get:
      tags: [Transactions]
      summary: List transaction requests
      parameters:
        - name: wallet_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TxStatus'
      responses:
        '200':
          description: List of transactions
  /v1/tx-requests/{tx_request_id}/approve:
    post:
      tags: [Transactions]
      summary: Approve/reject transaction
      description: SoD enforced - creator cannot approve
      parameters:
        - name: tx_request_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalCreate'
      responses:
        '200':
          description: Approval recorded
  /v1/cases:
    get:
      tags: [KYT Cases]
      summary: List KYT cases
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: direction
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of cases
  /v1/cases/{case_id}/resolve:
    post:
      tags: [KYT Cases]
      summary: Resolve KYT case
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KYTCaseResolve'
      responses:
        '200':
          description: Case resolved
...
  /v1/audit/packages/{tx_request_id}:
    get:
      tags: [Audit]
      summary: Get audit package for transaction
      parameters:
        - name: tx_request_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Audit package
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditPackageResponse'
  /v1/audit/verify:
    get:
      tags: [Audit]
      summary: Verify audit hash chain
      parameters:
        - name: from_sequence
          in: query
          schema:
            type: integer
        - name: to_sequence
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditVerifyResponse'
...